fix: main.py errors

+++ quick-start/main.py
@@ -8,8 +8,10 @@ import torch
 model_id = "TinyLlama/TinyLlama-1.1B-Chat-v1.0"
 
 config = CompilerConfig()
-model = NPUModelForCausalLM.from_pretrained(model_id, config, use_cache=True, dtype=torch.int8).eval()
-tokenizer = AutoTokenizer.from_pretrained(model_id, use_default_system_prompt=True)
+model = NPUModelForCausalLM.from_pretrained(
+    model_id, config, use_cache=True).eval()
+tokenizer = AutoTokenizer.from_pretrained(
+    model_id, use_default_system_prompt=True)
 tokenizer.pad_token_id = tokenizer.eos_token_id
 streamer = TextStreamer(tokenizer, skip_special_tokens=True)
